requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
apache-airflow-providers-http==5.3.4
requests==2.32.5
Flask-AppBuilder>=4.5.0,<5
apache-airflow-providers-amazon==9.15.0
minio==7.2.18